+++ src/runtime_src/driver/xclng/xrt/user_aws/shim.cpp
@@ -341,7 +341,7 @@ namespace awsbwhal {
       if (domain != XCL_MEM_DEVICE_RAM)
         return result;
 
-      unsigned ddr = 1;
+      uint64_t ddr = 1;
       ddr <<= flags;
       unsigned boHandle = xclAllocBO(size, XCL_BO_DEVICE_RAM, ddr);
       if (boHandle == mNullBO)
@@ -935,7 +935,7 @@ namespace awsbwhal {
 
     // Assume that the memory is always
     // created for the device ddr for now. Ignoring the flags as well.
-    unsigned int AwsXcl::xclAllocBO(size_t size, xclBOKind domain, unsigned flags)
+    unsigned int AwsXcl::xclAllocBO(size_t size, xclBOKind domain, uint64_t flags)
     {
       drm_xocl_create_bo info = {size, mNullBO, flags};
       int result = ioctl(mUserHandle, DRM_IOCTL_XOCL_CREATE_BO, &info);
@@ -945,7 +945,7 @@ namespace awsbwhal {
       return result ? mNullBO : info.handle;
     }
 
-    unsigned int AwsXcl::xclAllocUserPtrBO(void *userptr, size_t size, unsigned flags)
+    unsigned int AwsXcl::xclAllocUserPtrBO(void *userptr, size_t size, uint64_t flags)
     {
       drm_xocl_userptr_bo user = {reinterpret_cast<uint64_t>(userptr), size, mNullBO, flags};
       int result = ioctl(mUserHandle, DRM_IOCTL_XOCL_USERPTR_BO, &user);
@@ -1389,13 +1389,13 @@ int xclUnlockDevice(xclDeviceHandle handle)
   }
 }
 
-unsigned int xclAllocBO(xclDeviceHandle handle, size_t size, xclBOKind domain, unsigned flags)
+unsigned int xclAllocBO(xclDeviceHandle handle, size_t size, xclBOKind domain, uint64_t flags)
 {
   awsbwhal::AwsXcl *drv = awsbwhal::AwsXcl::handleCheck(handle);
   return drv ? drv->xclAllocBO(size, domain, flags) : -ENODEV;
 }
 
-unsigned int xclAllocUserPtrBO(xclDeviceHandle handle, void *userptr, size_t size, unsigned flags)
+unsigned int xclAllocUserPtrBO(xclDeviceHandle handle, void *userptr, size_t size, uint64_t flags)
 {
   awsbwhal::AwsXcl *drv = awsbwhal::AwsXcl::handleCheck(handle);
   return drv ? drv->xclAllocUserPtrBO(userptr, size, flags) : -ENODEV;
