+++ src/runtime_src/driver/xclng/xrt/user_aws/shim.h
@@ -169,8 +169,8 @@ public:
         //Sarab: Added for HAL2 XOCL Driver support
         //int xclGetErrorStatus(xclErrorStatus *info); Not supported for AWS
         bool xclUnlockDevice();
-        unsigned int xclAllocBO(size_t size, xclBOKind domain, unsigned flags);
-        unsigned int xclAllocUserPtrBO(void *userptr, size_t size, unsigned flags);
+        unsigned int xclAllocBO(size_t size, xclBOKind domain, uint64_t flags);
+        unsigned int xclAllocUserPtrBO(void *userptr, size_t size, uint64_t flags);
         void xclFreeBO(unsigned int boHandle);
         int xclWriteBO(unsigned int boHandle,
             const void *src, size_t size, size_t seek);
